# e2e/open-api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cats example",
    "description": "The cats API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "cats",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {}
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "api_key"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
      "ExtraModel": {
        "type": "object",
        "properties": {
          "one": {
            "type": "string"
          },
          "two": {
            "type": "number"
          }
        },
        "required": [
          "one",
          "two"
        ]
      },
      "BreadCoonDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "LettersEnum": {
        "type": "string",
        "enum": [
          "A",
          "B",
          "C"
        ]
      },
      "TagDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "name"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateCatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "age": {
            "type": "number",
            "minimum": 1,
            "maximum": 200
          },
          "_breed": {
            "discriminator": {
              "propertyName": "name",
              "mapping": {
                "MaineCoonDto": "#/components/schemas/MaineCoonDto",
                "BritishShorthairCatDto": "#/components/schemas/BritishShorthairCatDto"
              }
            },
            "allOf": [
              {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ]
              }
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "urls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "options": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "isReadonly": {
                  "type": "string"
                }
              }
            }
          },
          "enum": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C"
            ]
          },
          "enumArr": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "A",
                "B",
                "C"
              ]
            }
          },
          "tag": {
            "description": "tag",
            "allOf": [
              {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "name"
                  }
                },
                "required": [
                  "name"
                ]
              }
            ]
          }
        },
        "required": [
          "name",
          "age",
          "_breed",
          "tags",
          "createdAt",
          "urls",
          "options",
          "enum",
          "enumArr"
        ]
      },
      "Cat": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Kitty",
            "description": "The name of the Cat"
          },
          "age": {
            "type": "number",
            "example": 1,
            "minimum": 0,
            "description": "The age of the Cat"
          },
          "breed": {
            "type": "object",
            "example": "Maine Coon",
            "description": "The breed of the Cat"
          },
          "_tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "urls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "_options": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "isReadonly": {
                  "type": "string"
                }
              }
            }
          },
          "enum": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C"
            ]
          },
          "enumArr": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "A",
                "B",
                "C"
              ]
            }
          }
        },
        "required": [
          "name",
          "age",
          "breed",
          "_tags",
          "createdAt",
          "urls",
          "_options",
          "enum",
          "enumArr"
        ]
      },
      "MaineCoonDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "weight": {
            "type": "number"
          }
        },
        "required": [
          "name"
        ]
      },
      "BritishShorthairCatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "height": {
            "type": "number"
          }
        },
        "required": [
          "name"
        ]
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
    "/api/cats": {
      "post": {
        "operationId": "CatsController_create",
        "x-foo": {
          "test": "bar "
        },
        "summary": "Create cat",
        "parameters": [
          {
            "name": "header",
            "in": "header",
            "description": "Test",
            "required": false,
            "schema": {
              "default": "test",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "age": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 200
                  },
                  "_breed": {
                    "discriminator": {
                      "propertyName": "name",
                      "mapping": {
                        "MaineCoonDto": "#/components/schemas/MaineCoonDto",
                        "BritishShorthairCatDto": "#/components/schemas/BritishShorthairCatDto"
                      }
                    },
                    "allOf": [
                      {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "name"
                        ]
                      }
                    ]
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "createdAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "urls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "isReadonly": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "enum": {
                    "type": "string",
                    "enum": [
                      "A",
                      "B",
                      "C"
                    ]
                  },
                  "enumArr": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "A",
                        "B",
                        "C"
                      ]
                    }
                  },
                  "tag": {
                    "description": "tag",
                    "allOf": [
                      {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "description": "name"
                          }
                        },
                        "required": [
                          "name"
                        ]
                      }
                    ]
                  }
                },
                "required": [
                  "name",
                  "age",
                  "_breed",
                  "tags",
                  "createdAt",
                  "urls",
                  "options",
                  "enum",
                  "enumArr"
                ]
              },
              "examples": {
                "MyCat": {
                  "value": {
                    "name": "Yuumi",
                    "breed": {
                      "name": "British Shorthair",
                      "height": 99
                    }
                  }
                },
                "AnotherCat": {
                  "value": {
                    "name": "Kitty",
                    "breed": {
                      "name": "Maine Coon",
                      "weight": 10
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The record has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Kitty",
                      "description": "The name of the Cat"
                    },
                    "age": {
                      "type": "number",
                      "example": 1,
                      "minimum": 0,
                      "description": "The age of the Cat"
                    },
                    "breed": {
                      "type": "object",
                      "example": "Maine Coon",
                      "description": "The breed of the Cat"
                    },
                    "_tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "urls": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "_options": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "isReadonly": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "enum": {
                      "type": "string",
                      "enum": [
                        "A",
                        "B",
                        "C"
                      ]
                    },
                    "enumArr": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "A",
                          "B",
                          "C"
                        ]
                      }
                    }
                  },
                  "required": [
                    "name",
                    "age",
                    "breed",
                    "_tags",
                    "createdAt",
                    "urls",
                    "_options",
                    "enum",
                    "enumArr"
                  ]
                },
                "examples": {
                  "Full": {
                    "value": {
                      "name": "Yuumi"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "cats",
          "create cats"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "basic": []
          }
        ]
      },
      "get": {
        "operationId": "CatsController_findAll",
        "parameters": [
          {
            "name": "header",
            "in": "header",
            "description": "Test",
            "required": false,
            "schema": {
              "default": "test",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "exclusiveMaximum": true,
              "exclusiveMinimum": true,
              "title": "Page",
              "format": "int32",
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "_sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "enum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "A",
                "B",
                "C"
              ]
            }
          },
          {
            "name": "enumArr",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "A",
                  "B",
                  "C"
                ]
              }
            }
          },
          {
            "name": "beforeDate",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cats"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "basic": []
          }
        ]
      }
    },
    "/api/cats/{id}": {
      "get": {
        "operationId": "CatsController_findOne",
        "x-auth-type": "NONE",
        "parameters": [
          {
            "name": "header",
            "in": "header",
            "description": "Test",
            "required": false,
            "schema": {
              "default": "test",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The found record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Kitty",
                      "description": "The name of the Cat"
                    },
                    "age": {
                      "type": "number",
                      "example": 1,
                      "minimum": 0,
                      "description": "The age of the Cat"
                    },
                    "breed": {
                      "type": "object",
                      "example": "Maine Coon",
                      "description": "The breed of the Cat"
                    },
                    "_tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "urls": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "_options": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "isReadonly": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "enum": {
                      "type": "string",
                      "enum": [
                        "A",
                        "B",
                        "C"
                      ]
                    },
                    "enumArr": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "A",
                          "B",
                          "C"
                        ]
                      }
                    }
                  },
                  "required": [
                    "name",
                    "age",
                    "breed",
                    "_tags",
                    "createdAt",
                    "urls",
                    "_options",
                    "enum",
                    "enumArr"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "cats"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "basic": []
          }
        ]
      }
    },
    "/api/cats/explicit-query": {
      "get": {
        "operationId": "CatsController_findAllWithExplicitQuery",
        "parameters": [
          {
            "name": "header",
            "in": "header",
            "description": "Test",
            "required": false,
            "schema": {
              "default": "test",
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "exclusiveMaximum": true,
              "exclusiveMinimum": true,
              "title": "Page",
              "format": "int32",
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "_sortBy",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "enum",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "A",
                "B",
                "C"
              ]
            }
          },
          {
            "name": "enumArr",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "A",
                  "B",
                  "C"
                ]
              }
            }
          },
          {
            "name": "beforeDate",
            "in": "query",
            "required": true,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cats"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "basic": []
          }
        ]
      }
    },
    "/api/cats/bulk": {
      "get": {
        "operationId": "CatsController_findAllBulk",
        "parameters": [
          {
            "name": "header",
            "in": "header",
            "description": "Test",
            "required": false,
            "schema": {
              "default": "test",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cats"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "basic": []
          }
        ]
      },
      "post": {
        "operationId": "CatsController_createBulk",
        "parameters": [
          {
            "name": "header",
            "in": "header",
            "description": "Test",
            "required": false,
            "schema": {
              "default": "test",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cats"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "basic": []
          }
        ]
      }
    },
    "/api/cats/as-form-data": {
      "post": {
        "operationId": "CatsController_createAsFormData",
        "summary": "Create cat",
        "parameters": [
          {
            "name": "header",
            "in": "header",
            "description": "Test",
            "required": false,
            "schema": {
              "default": "test",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "age": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 200
                  },
                  "_breed": {
                    "discriminator": {
                      "propertyName": "name",
                      "mapping": {
                        "MaineCoonDto": "#/components/schemas/MaineCoonDto",
                        "BritishShorthairCatDto": "#/components/schemas/BritishShorthairCatDto"
                      }
                    },
                    "allOf": [
                      {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "name"
                        ]
                      }
                    ]
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "createdAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "urls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "isReadonly": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "enum": {
                    "type": "string",
                    "enum": [
                      "A",
                      "B",
                      "C"
                    ]
                  },
                  "enumArr": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "A",
                        "B",
                        "C"
                      ]
                    }
                  },
                  "tag": {
                    "description": "tag",
                    "allOf": [
                      {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "description": "name"
                          }
                        },
                        "required": [
                          "name"
                        ]
                      }
                    ]
                  }
                },
                "required": [
                  "name",
                  "age",
                  "_breed",
                  "tags",
                  "createdAt",
                  "urls",
                  "options",
                  "enum",
                  "enumArr"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The record has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Kitty",
                      "description": "The name of the Cat"
                    },
                    "age": {
                      "type": "number",
                      "example": 1,
                      "minimum": 0,
                      "description": "The age of the Cat"
                    },
                    "breed": {
                      "type": "object",
                      "example": "Maine Coon",
                      "description": "The breed of the Cat"
                    },
                    "_tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "urls": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "_options": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "isReadonly": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "enum": {
                      "type": "string",
                      "enum": [
                        "A",
                        "B",
                        "C"
                      ]
                    },
                    "enumArr": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "A",
                          "B",
                          "C"
                        ]
                      }
                    }
                  },
                  "required": [
                    "name",
                    "age",
                    "breed",
                    "_tags",
                    "createdAt",
                    "urls",
                    "_options",
                    "enum",
                    "enumArr"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "tags": [
          "cats"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "basic": []
          }
        ]
      }
    },
    "/api/cats/site*": {
      "get": {
        "operationId": "CatsController_getSite",
        "parameters": [
          {
            "name": "header",
            "in": "header",
            "description": "Test",
            "required": false,
            "schema": {
              "default": "test",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cats"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "basic": []
          }
        ]
      }
    }
  }
}